+++ test/test_udunits.cpp
@@ -94,7 +94,7 @@ TEST(UDUNITS, accepted_alias)
             auto nameUnit = units::unit_from_string(sname);
             if (is_error(nameUnit)) {
                 std::cout << "unable to convert " << sname
-                          << " into a valid unit \n";
+                          << " into a valid unit def=" << def << '\n';
                 ++failConvert;
             } else {
                 if (unit_cast(nameUnit) != unit_cast(definitionUnit)) {
@@ -114,7 +114,7 @@ TEST(UDUNITS, accepted_alias)
             auto symUnit = units::unit_from_string(symString);
             if (is_error(symUnit)) {
                 std::cout << "unable to convert " << symString
-                          << " into a valid unit \n";
+                          << " into a valid unit def=" << def << '\n';
                 ++failConvert;
             } else {
                 if (unit_cast(symUnit) != unit_cast(definitionUnit)) {
@@ -149,7 +149,8 @@ TEST(UDUNITS, derived_name_symbols)
 
         auto definitionUnit = units::unit_from_string(def);
         if (is_error(definitionUnit)) {
-            std::cout << "unable to convert " << def << " into a valid unit \n";
+            std::cout << "unable to convert " << def
+                      << " into a valid definition unit \n";
             ++failConvert;
         } else {
             auto name = cs->FirstChildElement("name");
@@ -159,7 +160,7 @@ TEST(UDUNITS, derived_name_symbols)
                 auto nameUnit = units::unit_from_string(sname);
                 if (is_error(nameUnit)) {
                     std::cout << "unable to convert " << sname
-                              << " into a valid unit \n";
+                              << " into a valid unit def=" << def << '\n';
                     ++failConvert;
                 } else {
                     if (unit_cast(nameUnit) != unit_cast(definitionUnit)) {
@@ -176,7 +177,7 @@ TEST(UDUNITS, derived_name_symbols)
                 auto symUnit = units::unit_from_string(symString);
                 if (is_error(symUnit)) {
                     std::cout << "unable to convert " << symString
-                              << " into a valid unit \n";
+                              << " into a valid unit def=" << def << '\n';
                     ++failConvert;
                 } else {
                     if (unit_cast(symUnit) != unit_cast(definitionUnit)) {
@@ -221,7 +222,7 @@ TEST(UDUNITS, derived_alias)
             auto nameUnit = units::unit_from_string(sname);
             if (is_error(nameUnit)) {
                 std::cout << "unable to convert " << sname
-                          << " into a valid unit \n";
+                          << " into a valid unit def=" << def << '\n';
                 ++failConvert;
             } else {
                 if (convert(unit_cast(nameUnit), unit_cast(definitionUnit)) !=
@@ -240,7 +241,7 @@ TEST(UDUNITS, derived_alias)
             auto symUnit = units::unit_from_string(symString);
             if (is_error(symUnit)) {
                 std::cout << "unable to convert " << symString
-                          << " into a valid unit \n";
+                          << " into a valid unit def=" << def << '\n';
                 ++failConvert;
             } else {
                 if (convert(unit_cast(symUnit), unit_cast(definitionUnit)) !=
@@ -282,7 +283,7 @@ TEST(UDUNITS, common_alias)
             auto nameUnit = units::unit_from_string(sname);
             if (is_error(nameUnit)) {
                 std::cout << "unable to convert " << sname
-                          << " into a valid unit \n";
+                          << " into a valid unit def=" << def << '\n';
                 ++failConvert;
             } else {
                 if (convert(unit_cast(nameUnit), unit_cast(definitionUnit)) !=
@@ -319,7 +320,7 @@ TEST(UDUNITS, common_alias)
             auto symUnit = units::unit_from_string(symString);
             if (is_error(symUnit)) {
                 std::cout << "unable to convert " << symString
-                          << " into a valid unit \n";
+                          << " into a valid unit def=" << def << '\n';
                 ++failConvert;
             } else {
                 if (convert(unit_cast(symUnit), unit_cast(definitionUnit)) !=